refactor: consolidate inventory controller and routes
